+++ src/jsMain/kotlin/at/posselt/pfrpg2e/migrations/migrations/Migration20.kt
@@ -63,5 +63,6 @@ class Migration20 : Migration(20) {
             .toMutableRecord()
         camping.forcedMarchActive = false
         camping.secondsSpentForcedMarching = 0
+        camping.hexSizeInMiles = 12
     }
 }